Revert "Dev cy (#782)"
This reverts commit afc7f6987045a85e099ecde17b5c885b11768ae3.
