+++ components.d.ts
@@ -29,7 +29,6 @@ declare module 'vue' {
     copy: typeof import('./src/components/Global/Provider copy.vue')['default']
     CopyLyrics: typeof import('./src/components/Modal/CopyLyrics.vue')['default']
     CoverList: typeof import('./src/components/List/CoverList.vue')['default']
-    CoverManager: typeof import('./src/components/Modal/Setting/CoverManager.vue')['default']
     CoverMenu: typeof import('./src/components/Menu/CoverMenu.vue')['default']
     CreatePlaylist: typeof import('./src/components/Modal/CreatePlaylist.vue')['default']
     CustomCode: typeof import('./src/components/Modal/Setting/CustomCode.vue')['default']
@@ -162,7 +161,6 @@ declare module 'vue' {
     PlayerSlider: typeof import('./src/components/Player/PlayerComponents/PlayerSlider.vue')['default']
     PlayerSpectrum: typeof import('./src/components/Player/PlayerComponents/PlayerSpectrum.vue')['default']
     PlaylistAdd: typeof import('./src/components/Modal/PlaylistAdd.vue')['default']
-    PlaylistPageManager: typeof import('./src/components/Modal/Setting/PlaylistPageManager.vue')['default']
     PlaySetting: typeof import('./src/components/Setting/old/PlaySetting.vue')['default']
     Provider: typeof import('./src/components/Global/Provider.vue')['default']
     ProxyConfig: typeof import('./src/components/Setting/components/ProxyConfig.vue')['default']
