+++ src/components/Player/MainPlayer.vue
@@ -16,7 +16,6 @@
       <!-- 封面 -->
       <Transition name="fade" mode="out-in">
         <div
-          v-if="!settingStore.hiddenCovers.player"
           :key="musicStore.playSong.cover"
           class="cover"
           @click.stop="statusStore.showFullPlayer = true"
@@ -46,16 +45,12 @@
             <TextContainer
               :key="musicStore.playSong.name"
               :text="
-                settingStore.hideBracketedContent
+                settingStore.hideLyricBrackets
                   ? removeBrackets(musicStore.playSong.name)
                   : musicStore.playSong.name
               "
               :speed="0.2"
               class="name"
-              style="cursor: pointer"
-              @click.stop="
-                settingStore.hiddenCovers.player && (statusStore.showFullPlayer = true)
-              "
             />
             <!-- 倍速 -->
             <n-tag
@@ -102,21 +97,11 @@
                   class="ar-item"
                   @click="openJumpArtist(musicStore.playSong.artists, item.id)"
                 >
-                  {{
-                    settingStore.hideBracketedContent ? removeBrackets(item.name) : item.name
-                  }}
+                  {{ item.name }}
                 </n-text>
               </template>
-              <n-text
-                v-else
-                class="ar-item"
-                @click="openJumpArtist(musicStore.playSong.artists)"
-              >
-                {{
-                  settingStore.hideBracketedContent
-                    ? removeBrackets(musicStore.playSong.artists)
-                    : musicStore.playSong.artists || "未知艺术家"
-                }}
+              <n-text v-else class="ar-item" @click="openJumpArtist(musicStore.playSong.artists)">
+                {{ musicStore.playSong.artists || "未知艺术家" }}
               </n-text>
             </div>
           </Transition>
