+++ src/components/Player/PlayerComponents/PlayerComment.vue
@@ -17,7 +17,7 @@
       </n-image>
       <n-flex :size="2" class="song-info" vertical>
         <span class="title text-hidden">{{
-          settingStore.hideBracketedContent
+          settingStore.hideLyricBrackets
             ? removeBrackets(musicStore.playSong.name)
             : musicStore.playSong.name
         }}</span>
