+++ src/core/player/LyricManager.ts
@@ -677,195 +677,6 @@ class LyricManager {
   }
 
   /**
-   * 替换歌词括号内容
-   * @param lyricData 歌词数据
-   * @returns 替换后的歌词数据
-   */
-  private applyBracketReplacement(lyricData: SongLyric): SongLyric {
-    const settingStore = useSettingStore();
-    if (!settingStore.replaceLyricBrackets) {
-      return lyricData;
-    }
-
-    const newLyricData = cloneDeep(lyricData);
-
-    // --- Configuration Helper ---
-    // Helper to determine the replacement strategy based on settings
-    const getReplacementConfig = () => {
-      const preset = settingStore.bracketReplacementPreset || "dash";
-      const custom = settingStore.customBracketReplacement || "-";
-      let startStr = " - ";
-      let endStr = " ";
-      let isEnclosure = false;
-
-      if (preset === "angleBrackets") {
-        startStr = "〔";
-        endStr = "〕";
-        isEnclosure = true;
-      } else if (preset === "cornerBrackets") {
-        startStr = "「";
-        endStr = "」";
-        isEnclosure = true;
-      } else if (preset === "custom") {
-        const trimmed = custom.trim();
-        // Heuristic: if length is 2 and not just dashes, treat as pair (e.g. "()")
-        // This allows users to input "()" to mean replace ( with ( and ) with )
-        if (trimmed.length === 2 && trimmed[0] !== trimmed[1] && !trimmed.includes("-")) {
-          startStr = trimmed[0];
-          endStr = trimmed[1];
-          isEnclosure = true;
-        } else {
-          // Treat as separator
-          startStr = " " + trimmed + " ";
-          startStr = startStr.replace(/\s+/g, " "); // Normalize spaces
-          endStr = " ";
-          isEnclosure = false;
-        }
-      }
-      return { startStr, endStr, isEnclosure };
-    };
-
-    const { startStr, endStr, isEnclosure } = getReplacementConfig();
-
-    // --- Processing Helpers ---
-
-    // Process a string (used for translations and romaji)
-    const processString = (str: string): string => {
-      if (!str) return str;
-
-      // If the entire string is enclosed in brackets (e.g. "(Chorus)"), remove them if not in enclosure mode
-      if (!isEnclosure && /^\s*[\(（][^()（）]*[\)）]\s*$/.test(str)) {
-        return str
-          .replace(/^\s*[\(（]/, "")
-          .replace(/[\)）]\s*$/, "")
-          .trim();
-      }
-
-      let res = str.replace(/[\(（]/g, startStr);
-      if (isEnclosure) {
-        res = res.replace(/[\)）]/g, endStr);
-      } else {
-        // Separator mode:
-        // 1. Remove ) if it's at the end of the string (effectively just a closing marker)
-        // 2. Otherwise replace ) with endStr (usually space)
-        res = res
-          .replace(/[\)）](?=\s*$)/g, "")
-          .replace(/[\)）]/g, endStr);
-
-        // Cleanup double dashes if the separator contains a dash
-        if (startStr.includes("-")) {
-          res = res.replace(/(?:\s*-\s*){2,}/g, " - ");
-        }
-      }
-      return res;
-    };
-
-    // Process a single lyric line (LRC/YRC)
-    const processLine = (line: LyricLine) => {
-      // 1. Check for "Full Bracket" line (e.g. "(Music)")
-      // If the whole line is in brackets and we are NOT in enclosure mode (e.g. dash mode),
-      // we likely want to strip the brackets entirely instead of replacing them with dashes.
-      const fullText = line.words.map((w) => w.word).join("");
-      const isFullBracket = /^\s*[\(（][^()（）]*[\)）]\s*$/.test(fullText);
-
-      if (isFullBracket && !isEnclosure) {
-        // Remove the first opening bracket found in the words
-        let foundStart = false;
-        for (const word of line.words) {
-          if (foundStart) break;
-          if (/[\(（]/.test(word.word)) {
-            word.word = word.word.replace(/[\(（]/, "");
-            foundStart = true;
-          }
-        }
-        // Remove the last closing bracket found in the words
-        let foundEnd = false;
-        for (let i = line.words.length - 1; i >= 0; i--) {
-          if (foundEnd) break;
-          const word = line.words[i];
-          if (/[\)）]/.test(word.word)) {
-            // Find the last occurrence of ) or ）
-            const lastIndex = Math.max(word.word.lastIndexOf(")"), word.word.lastIndexOf("）"));
-            if (lastIndex !== -1) {
-              word.word =
-                word.word.substring(0, lastIndex) + word.word.substring(lastIndex + 1);
-              foundEnd = true;
-            }
-          }
-        }
-      } else {
-        // Normal replacement logic
-        line.words.forEach((word, index) => {
-          // Replace opening brackets
-          word.word = word.word.replace(/[\(（]/g, startStr);
-
-          if (isEnclosure) {
-            // Enclosure mode: simply replace closing brackets with endStr
-            word.word = word.word.replace(/[\)）]/g, endStr);
-          } else {
-            // Separator mode: logic to handle closing brackets nicely
-            word.word = word.word.replace(/[\)）]/g, (_, offset, string) => {
-              const isAtEnd = offset === string.length - 1;
-              // If ) is at the end of the word...
-              if (isAtEnd) {
-                // ...and it's the last word of the line, remove it (it's just closing the line)
-                if (index === line.words.length - 1) {
-                  return "";
-                } else {
-                  // ...otherwise it's a separator between this word and the next
-                  return endStr;
-                }
-              } else {
-                // If not at end of word, it's a separator
-                return endStr;
-              }
-            });
-          }
-        });
-
-        // Cleanup double dashes (only for separator mode with dash)
-        if (!isEnclosure && startStr.includes("-")) {
-          line.words.forEach((word, index) => {
-            // 1. Intra-word cleanup: " -  - " -> " - "
-            word.word = word.word.replace(/(?:\s*-\s*){2,}/g, " - ");
-
-            // 2. Inter-word cleanup: Prev word ends with dash, current starts with dash
-            if (index > 0) {
-              const prev = line.words[index - 1];
-              if (/-\s*$/.test(prev.word) && /^\s*-/.test(word.word)) {
-                // Remove trailing dash from previous word
-                prev.word = prev.word.replace(/-\s*$/, "");
-                // Ensure current word starts with proper separator
-                if (!/^\s*-\s+/.test(word.word)) {
-                  word.word = " - " + word.word.replace(/^\s*-\s*/, "");
-                }
-              }
-            }
-          });
-        }
-      }
-
-      // Process extra fields
-      if (line.translatedLyric) {
-        line.translatedLyric = processString(line.translatedLyric);
-      }
-      if (line.romanLyric) {
-        line.romanLyric = processString(line.romanLyric);
-      }
-    };
-
-    const processLines = (lines: LyricLine[] | undefined) => {
-      if (!lines) return;
-      lines.forEach(processLine);
-    };
-
-    processLines(newLyricData.lrcData);
-    processLines(newLyricData.yrcData);
-
-    return newLyricData;
-  }
-
-  /**
    * 比较歌词数据是否相同
    * @param oldData 旧歌词数据
    * @param newData 新歌词数据
@@ -927,9 +738,6 @@ class LyricManager {
     const statusStore = useStatusStore();
     // 若非本次
     if (this.activeLyricReq !== req) return;
-
-    // 应用括号替换
-    lyricData = this.applyBracketReplacement(lyricData);
 
     // 规范化时间
     this.normalizeLyricLines(lyricData.yrcData);
