+++ src/utils/modal.ts
@@ -68,22 +68,6 @@ export const openUserAgreement = async () => {
   });
 };
 
-/** 打开歌单界面配置弹窗 */
-export const openPlaylistPageManager = async () => {
-  const { default: PlaylistPageManager } =
-    await import("@/components/Modal/Setting/PlaylistPageManager.vue");
-  window.$modal.create({
-    preset: "card",
-    transformOrigin: "center",
-    autoFocus: false,
-    style: { width: "500px" },
-    title: "歌单界面配置",
-    content: () => {
-      return h(PlaylistPageManager);
-    },
-  });
-};
-
 // 用户登录
 export const openUserLogin = async (
   showTip: boolean = false,
@@ -426,22 +410,6 @@ export const openSidebarHideManager = async () => {
     title: "侧边栏隐藏管理",
     content: () => {
       return h(SidebarHideManager);
-    },
-  });
-};
-
-/** 打开封面隐藏配置弹窗 */
-export const openCoverManager = async () => {
-  const { default: CoverManager } =
-    await import("@/components/Modal/Setting/CoverManager.vue");
-  window.$modal.create({
-    preset: "card",
-    transformOrigin: "center",
-    autoFocus: false,
-    style: { width: "500px" },
-    title: "封面隐藏配置",
-    content: () => {
-      return h(CoverManager);
     },
   });
 };
