+++ src/views/Discover/artists.vue
@@ -28,13 +28,7 @@
         </n-tag>
       </n-flex>
     </div>
-    <ArtistList
-      :data="artistsData"
-      :loading="loading"
-      :loadMore="hasMore"
-      :hiddenCover="settingStore.hiddenCovers.artist"
-      @loadMore="loadMore"
-    />
+    <ArtistList :data="artistsData" :loading="loading" :loadMore="hasMore" @loadMore="loadMore" />
   </div>
 </template>
 
@@ -42,10 +36,8 @@
 import type { ArtistType } from "@/types/main";
 import { artistTypeList } from "@/api/artist";
 import { formatArtistsList } from "@/utils/format";
-import { useSettingStore } from "@/stores";
 
 const router = useRouter();
-const settingStore = useSettingStore();
 
 // 歌手标签数据
 const artistInitials = [
