+++ src/views/Discover/toplists.vue
@@ -11,7 +11,6 @@
               :height="160"
               :description="item.updateTip"
               size="normal"
-              :hiddenCover="settingStore.hiddenCovers.toplist"
               @click="router.push({ name: 'playlist', query: { id: item.id } })"
             >
               <template #info>
@@ -31,8 +30,8 @@
       <div v-else class="official-list">
         <n-grid cols="1 600:2 1000:3" x-gap="20" y-gap="20">
           <n-gi v-for="item in 4" :key="item">
-            <n-card class="loading" :class="{ 'no-cover': settingStore.hiddenCovers.toplist }">
-              <n-skeleton v-if="!settingStore.hiddenCovers.toplist" class="cover" />
+            <n-card class="loading">
+              <n-skeleton class="cover" />
               <div class="desc">
                 <n-skeleton text round :repeat="3" />
               </div>
@@ -42,12 +41,7 @@
       </div>
     </Transition>
     <n-divider style="margin-bottom: 0"> 精选榜 </n-divider>
-    <CoverList
-      :data="topListData.selected"
-      :loading="true"
-      type="playlist"
-      :hiddenCover="settingStore.hiddenCovers.toplist"
-    />
+    <CoverList :data="topListData.selected" :loading="true" type="playlist" />
   </div>
 </template>
 
@@ -55,10 +49,8 @@
 import { topPlaylist } from "@/api/playlist";
 import type { CoverType } from "@/types/main";
 import { formatCoverList } from "@/utils/format";
-import { useSettingStore } from "@/stores";
 
 const router = useRouter();
-const settingStore = useSettingStore();
 
 // 排行榜数据
 const topListData = ref<{
@@ -114,15 +106,6 @@ onMounted(getTopPlaylistData);
       width: 100%;
       :deep(.n-skeleton) {
         height: 20px;
-      }
-    }
-    &.no-cover {
-      :deep(.n-card__content) {
-        padding: 12px;
-      }
-      .desc {
-        justify-content: center;
-        gap: 12px;
       }
     }
   }
