+++ src/views/Like/albums.vue
@@ -1,17 +1,11 @@
 <template>
   <div class="like-type">
-    <CoverList
-      :data="dataStore.userLikeData.albums"
-      :loading="true"
-      type="album"
-      :hiddenCover="settingStore.hiddenCovers.like"
-    />
+    <CoverList :data="dataStore.userLikeData.albums" :loading="true" type="album" />
   </div>
 </template>
 
 <script setup lang="ts">
-import { useDataStore, useSettingStore } from "@/stores";
+import { useDataStore } from "@/stores";
 
 const dataStore = useDataStore();
-const settingStore = useSettingStore();
 </script>
