+++ src/views/Like/artists.vue
@@ -1,16 +1,11 @@
 <template>
   <div class="like-type">
-    <ArtistList
-      :data="dataStore.userLikeData.artists"
-      :loading="true"
-      :hiddenCover="settingStore.hiddenCovers.like"
-    />
+    <ArtistList :data="dataStore.userLikeData.artists" :loading="true" />
   </div>
 </template>
 
 <script setup lang="ts">
-import { useDataStore, useSettingStore } from "@/stores";
+import { useDataStore } from "@/stores";
 
 const dataStore = useDataStore();
-const settingStore = useSettingStore();
 </script>
