+++ src/views/Like/radios.vue
@@ -1,17 +1,11 @@
 <template>
   <div class="like-type">
-    <CoverList
-      :data="dataStore.userLikeData.djs"
-      :loading="true"
-      type="radio"
-      :hiddenCover="settingStore.hiddenCovers.like"
-    />
+    <CoverList :data="dataStore.userLikeData.djs" :loading="true" type="radio" />
   </div>
 </template>
 
 <script setup lang="ts">
-import { useDataStore, useSettingStore } from "@/stores";
+import { useDataStore } from "@/stores";
 
 const dataStore = useDataStore();
-const settingStore = useSettingStore();
 </script>
