+++ src/views/Like/videos.vue
@@ -5,14 +5,12 @@
       :loading="true"
       cols="2 600:2 800:3 900:4 1200:5 1400:6"
       type="video"
-      :hiddenCover="settingStore.hiddenCovers.like"
     />
   </div>
 </template>
 
 <script setup lang="ts">
-import { useDataStore, useSettingStore } from "@/stores";
+import { useDataStore } from "@/stores";
 
 const dataStore = useDataStore();
-const settingStore = useSettingStore();
 </script>
